[dsbx] refactor: Live* introspection types to types/db.ts; errorEnvelope takes the typed error (review)
LiveColumn/LiveIndex/LiveTable are pure introspection shapes — they move
to the dependency-free types/db.ts alongside the rest of the db contract.
errorEnvelope now takes a DbCommandError (the Err branch of a Result)
instead of sniffing unknown with instanceof: expected errors are typed
values, so no runtime type check is needed at the serialization boundary.
